+++ source/patterns/@aws-solutions-constructs/aws-events-rule-sqs/test/events-rule-sqs-queue.test.ts
@@ -12,22 +12,22 @@
  */
 
 import * as cdk from '@aws-cdk/core';
-import { EventsRuleToSQS, EventsRuleToSQSProps } from '../lib';
+import { EventsRuleToSqs, EventsRuleToSqsProps } from '../lib';
 import * as events from "@aws-cdk/aws-events";
 import { SynthUtils } from '@aws-cdk/assert';
 import '@aws-cdk/assert/jest';
 import * as defaults from '@aws-solutions-constructs/core';
 
 function deployNewStack(stack: cdk.Stack) {
-  const props: EventsRuleToSQSProps = {
+  const props: EventsRuleToSqsProps = {
       eventRuleProps: {
           schedule: events.Schedule.rate(cdk.Duration.minutes(5))
       }
   };
-  return new EventsRuleToSQS(stack, 'test-events-rule-sqs', props);
+  return new EventsRuleToSqs(stack, 'test-events-rule-sqs', props);
 }
 
-test('snapshot test EventsRuleToSQS default params', () => {
+test('snapshot test EventsRuleToSqs default params', () => {
   const stack = new cdk.Stack();
   deployNewStack(stack);
   expect(SynthUtils.toCloudFormation(stack)).toMatchSnapshot();
@@ -61,14 +61,14 @@ test('check the sqs queue properties with existing KMS key', () => {
     description: 'my-key'
   });
 
-  const props: EventsRuleToSQSProps = {
+  const props: EventsRuleToSqsProps = {
     eventRuleProps: {
         schedule: events.Schedule.rate(cdk.Duration.minutes(5))
     },
     encryptionKey: key
   };
 
-  new EventsRuleToSQS(stack, 'test-events-rule-sqs', props);
+  new EventsRuleToSqs(stack, 'test-events-rule-sqs', props);
 
   expect(stack).toHaveResource('AWS::SQS::Queue', {
     KmsMasterKeyId: {
@@ -312,7 +312,7 @@ test('check if the dead letter queue policy is setup', () => {
 
 test('check properties', () => {
   const stack = new cdk.Stack();
-  const construct: EventsRuleToSQS = deployNewStack(stack);
+  const construct: EventsRuleToSqs = deployNewStack(stack);
 
   expect(construct.eventsRule !== null);
   expect(construct.sqsQueue !== null);